+++ spec/models/bigbluebutton_recording_db_spec.rb
@@ -10,8 +10,8 @@ describe BigbluebuttonRecording do
     it { should have_db_column(:meetingid).of_type(:string) }
     it { should have_db_column(:name).of_type(:string) }
     it { should have_db_column(:published).of_type(:boolean) }
-    it { should have_db_column(:start_time).of_type(:datetime) }
-    it { should have_db_column(:end_time).of_type(:datetime) }
+    it { should have_db_column(:start_time).of_type(:integer) }
+    it { should have_db_column(:end_time).of_type(:integer) }
     it { should have_db_column(:available).of_type(:boolean) }
     it { should have_db_column(:description).of_type(:string) }
     it { should have_db_column(:created_at).of_type(:datetime) }