+++ spec/models/bigbluebutton_room_spec.rb
@@ -64,7 +64,7 @@ describe BigbluebuttonRoom do
 
   # attr_accessors
   [:running, :participant_count, :moderator_count, :attendees,
-   :has_been_forcibly_ended, :start_time, :end_time, :external,
+   :has_been_forcibly_ended, :create_time, :end_time, :external,
    :request_headers, :record_meeting, :duration].each do |attr|
     it { should respond_to(attr) }
     it { should respond_to("#{attr}=") }
@@ -136,7 +136,7 @@ describe BigbluebuttonRoom do
       room.moderator_count.should be(0)
       room.running.should be(false)
       room.has_been_forcibly_ended.should be(false)
-      room.start_time.should be_nil
+      room.create_time.should be_nil
       room.end_time.should be_nil
       room.attendees.should eql([])
       room.request_headers.should == {}
@@ -308,7 +308,7 @@ describe BigbluebuttonRoom do
         it { room.has_been_forcibly_ended.should == false }
         it { room.participant_count.should == 0 }
         it { room.moderator_count.should == 0 }
-        it { room.start_time.should == nil }
+        it { room.create_time.should == nil }
         it { room.reload.create_time.should == nil }
         it { room.end_time.should == nil }
         it { room.attendees.should == [] }
@@ -325,7 +325,6 @@ describe BigbluebuttonRoom do
         it { room.has_been_forcibly_ended.should == false }
         it { room.participant_count.should == 4 }
         it { room.moderator_count.should == 2 }
-        it { room.start_time.should == DateTime.parse("Wed Apr 06 17:09:57 UTC 2011") }
         it { room.reload.create_time.should == 1466876909 }
         it { room.end_time.should == nil }
         it {
@@ -642,8 +641,6 @@ describe BigbluebuttonRoom do
           it { subject.name.should eql(room.name) }
           it { subject.recorded.should eql(room.record_meeting) }
           it { subject.create_time.should eql(room.create_time) }
-          it { subject.create_time.should eql(room.create_time) }
-          it { subject.start_time.should eql(room.start_time) }
           it { subject.ended.should eql(false) }
         end
 
@@ -1243,8 +1240,8 @@ describe BigbluebuttonRoom do
   describe "#get_current_meeting" do
     let(:room) { FactoryGirl.create(:bigbluebutton_room) }
 
-    context "if there's no start_time set in the room" do
-      before { room.start_time = nil }
+    context "if there's no create_time set in the room" do
+      before { room.create_time = nil }
       it { room.get_current_meeting.should be_nil }
     end
 
@@ -1280,7 +1277,7 @@ describe BigbluebuttonRoom do
       before {
         room.create_time = Time.now.utc
         room.running = !room.running # to change its default value
-        room.start_time = Time.at(Time.now.to_i - 123)  # to change its default value
+        room.create_time = Time.at(Time.now.to_i - 123)  # to change its default value
       }
 
       context "and no metadata was passed" do
@@ -1294,7 +1291,7 @@ describe BigbluebuttonRoom do
         }
         subject { BigbluebuttonMeeting.find_by_room_id(room.id) }
         it("sets running") { subject.running.should eq(room.running) }
-        it("sets start_time") { subject.start_time.utc.to_i.should eq(room.start_time.utc.to_i) }
+        it("sets create_time") { subject.create_time.should eq(room.create_time.to_i) }
         it("doesn't set creator_id") { subject.creator_id.should be_nil }
         it("doesn't set creator_name") { subject.creator_name.should be_nil }
       end
@@ -1310,7 +1307,7 @@ describe BigbluebuttonRoom do
         }
         subject { BigbluebuttonMeeting.find_by_room_id(room.id) }
         it("sets running") { subject.running.should eq(room.running) }
-        it("sets start_time") { subject.start_time.utc.to_i.should eq(room.start_time.utc.to_i) }
+        it("sets create_time") { subject.create_time.should eq(room.create_time.to_i) }
         it("sets creator_id") { subject.creator_id.should eq(user.id) }
         it("sets creator_name") { subject.creator_name.should eq(user.name) }
       end
@@ -1326,7 +1323,7 @@ describe BigbluebuttonRoom do
         }
         subject { BigbluebuttonMeeting.find_by_room_id(room.id) }
         it("sets running") { subject.running.should eq(room.running) }
-        it("sets start_time") { subject.start_time.utc.to_i.should eq(room.start_time.utc.to_i) }
+        it("sets create_time") { subject.create_time.should eq(room.create_time.to_i) }
         it("sets ended") { subject.ended.should be(false) }
       end
     end
@@ -1340,7 +1337,7 @@ describe BigbluebuttonRoom do
       let!(:meeting) { FactoryGirl.create(:bigbluebutton_meeting, room: room, ended: false, running: true, create_time: room.create_time) }
       subject {
         expect {
-          room.create_meeting_record
+          room.create_meeting_record({}, server)
         }.not_to change{ BigbluebuttonMeeting.count }
       }
       it { BigbluebuttonMeeting.where(room: room).count.should be(1) }
@@ -1368,15 +1365,14 @@ describe BigbluebuttonRoom do
         room.create_time = Time.now.utc
         room.running = !room.running # to change its default value
         room.record_meeting = !room.record_meeting # to change its default value
-        room.start_time = Time.at(Time.now.to_i - 123)  # to change its default value
+        room.create_time = Time.at(Time.now.to_i - 123)  # to change its default value
       }
 
       context "if there's no meeting associated yet creates one" do
-        context "and no metadata was passed" do
+        context "and there's no metadata in the response" do
           before(:each) {
             expect {
-              room.should_receive(:select_server).and_return(server)
-              room.create_meeting_record
+              room.create_meeting_record({}, server)
             }.to change{ BigbluebuttonMeeting.count }.by(1)
           }
           subject { BigbluebuttonMeeting.last }
@@ -1387,20 +1383,40 @@ describe BigbluebuttonRoom do
           it("sets name") { subject.name.should eq(room.name) }
           it("sets recorded") { subject.recorded.should eq(room.record_meeting) }
           it("sets running") { subject.running.should eq(room.running) }
-          it("sets start_time") { subject.start_time.utc.to_i.should eq(room.start_time.utc.to_i) }
+          it("sets create_time") { subject.create_time.should eq(room.create_time.to_i) }
           it("doesn't set creator_id") { subject.creator_id.should be_nil }
           it("doesn't set creator_name") { subject.creator_name.should be_nil }
         end
 
-        context "and metadata was passed" do
+        context "and there's metadata in the response" do
           before(:each) {
             expect {
-              room.create_meeting_record(metadata)
+              room.create_meeting_record({ metadata: metadata }, server)
             }.to change{ BigbluebuttonMeeting.count }.by(1)
           }
           subject { BigbluebuttonMeeting.last }
           it("sets creator_id") { subject.creator_id.should eq(user.id) }
           it("sets creator_name") { subject.creator_name.should eq(user.name) }
+        end
+
+        context "and there are user attributes" do
+          let(:user_attrs) {
+            {
+              meetingID: room.meetingid + "-2",
+              name: room.name + "-2",
+              record: false # important to be false here
+            }
+          }
+          before(:each) {
+            room.record_meeting = true
+            expect {
+              room.create_meeting_record({}, server, user_attrs)
+            }.to change{ BigbluebuttonMeeting.count }.by(1)
+          }
+          subject { BigbluebuttonMeeting.last }
+          it("sets meetingid") { subject.meetingid.should eq(room.meetingid + '-2') }
+          it("sets name") { subject.name.should eq(room.name + '-2') }
+          it("sets recorded") { subject.recorded.should eq(false) }
         end
       end
 
@@ -1411,7 +1427,7 @@ describe BigbluebuttonRoom do
         before(:each) {
           BigbluebuttonMeeting.where(room: room, ended: false).count.should be(2)
           expect {
-            room.create_meeting_record(metadata)
+            room.create_meeting_record({ metadata: metadata }, server)
           }.to change{ BigbluebuttonMeeting.count }.by(1)
         }
         it { BigbluebuttonMeeting.where(room: room).count.should be(3) }
@@ -1455,6 +1471,26 @@ describe BigbluebuttonRoom do
       it { meeting.reload.running.should be(false) }
       it { meeting.reload.ended.should be(true) }
     end
+
+    context "enqueues a worker to fetch recordings" do
+
+      context "if at least one meeting was ended" do
+        let!(:meeting1) { FactoryGirl.create(:bigbluebutton_meeting, room: room, ended: false, running: true) }
+        before {
+          expect(Resque).to receive(:enqueue_in).with(4.minutes, ::BigbluebuttonRecordingsForRoom, room.id, 3)
+        }
+        it { room.finish_meetings }
+      end
+
+      context "not if no meeting was ended" do
+        let!(:meeting1) { FactoryGirl.create(:bigbluebutton_meeting, room: room, ended: true, running: true) }
+        before {
+          expect(Resque).not_to receive(:enqueue_in)
+        }
+        it { room.finish_meetings }
+      end
+
+    end
   end
 
   describe "#internal_create_meeting" do
@@ -1464,6 +1500,7 @@ describe BigbluebuttonRoom do
     it "sets the request headers"
     it "calls api.create_meeting"
     it "accepts additional user options to override the options in the database"
+    it "returns the server selected and the response"
 
     context "adds the invitation URL, if any" do
       before { mock_server_and_api }